// src/app/core/models/producto.model.ts
export interface Producto {
  id: string;
  descripcion: string;
  categoria: string;
  imagen: string;
  preciocompra: string;
  preciounitario: string;
  codigo: string;
  sku: string;
  medida: string;
  tipoafetacion: string;
  adddate: string;
  idusercreate: string;
  editdate: string;
  iduserupdate: string;
  status: string;
  sede: string;
  cantidad: string;
}